Enhancement: Revamp question card ui
